Pull Request - 강의, 콘텐츠 기능 구현

+++ src/main/java/com/example/onlineeduplatformlecture/service/LectureService.java
@@ -12,7 +12,7 @@ public interface LectureService {
     Flux<Lecture> getLectureList();
 
     // GET /lectures/{lectureId} (강의 열람)
-    Mono<Lecture> getLecture(int lectureId);
+    Mono<Lecture> getLecture(Long lectureId);
 
 
 }
